app/video_parse/requirements.txt
fastapi==0.115.11
uvicorn==0.23.1
httpx==0.27.0
mcp==1.3.0
openapi-httpx-client==0.3.1
httmcp==0.1.4
nest-asyncio
python-dotenv
